feat(content-studio): lock modular editor flows with smoke tests
Split content-studio into focused modules and add Playwright smoke coverage so core authoring and lint navigation flows stay reproducible after the refactor.
